😸refak: add config.js
