+++ src/config.js
@@ -1,0 +1,11 @@
+import theme from "@/themes"
+
+export const seatColorMapping = {
+    FC_EXPECTED_WIN_DEMO: theme.palette.warning.main,
+    GC_EXPECTED_WIN_DEMO: theme.palette.warning.light,
+    GC_EXPECTED_WIN_OTHER: theme.palette.success.main,
+    FC_EXPECTED_WIN_OTHER: theme.palette.success.main,
+    UNRESOLVED: theme.palette.divider,
+    GC_EXPECTED_WIN_BEIJING: theme.palette.info.light,
+    FC_EXPECTED_WIN_BEIJING: theme.palette.info.main,
+  }
